Update README.md
How can we retrieve the book title, category, and rental price of all available books?
What are the employee names and their respective salaries in descending order of salary?
How can we retrieve the book titles and corresponding customers who have issued those books?
What is the total count of books in each category?
How can we retrieve the employee names and positions for employees with salaries above Rs. 50,000?
Which customer names registered before 2022-01-01 and have not issued any books yet?
What are the branch numbers and the total count of employees in each branch?
How can we retrieve the names of customers who have issued books in the month of June 2023?
How can we retrieve book titles from the "Books" table containing the category "History"?
What are the branch numbers and employee counts for branches that have more than five employees?